RHDEVDOCS-3565: Fix API reference on master (#121)
This commit primarily modifies the `.gitignore` file to include the
locally-generated API reference for the in-development version. This
commit also introduces the generated reference to Git and ensures that
we ignore files ending in `.css.map`. These files are a side effect of
the rendering process and are not necessary for rendering content.